+++ CloudShell.Abstractions/Logging/CloudShellLogCategories.cs
@@ -11,6 +11,7 @@ public static class CloudShellLogCategories
     public const string ResourceHealthProbeHttpClient = "System.Net.Http.HttpClient." + ResourceHealthProbes;
     public const string ResourceRecovery = ResourceManager + ".ResourceRecovery";
     public const string ResourceRecoveryPolling = ResourceRecovery + ".Polling";
+    public const string ResourceReplicaManagement = ResourceManager + ".ReplicaManagement";
     public const string ProgrammaticResourceStartup = ResourceManager + ".ProgrammaticResourceStartup";
     public const string HostScopedResourceShutdown = ResourceManager + ".HostScopedResourceShutdown";
     public const string ResourceLifecycle = ResourceManager + ".Lifecycle";